% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configClasses.R
\docType{class}
\name{CATConfig-class}
\alias{CATConfig-class}
\alias{CATConfig}
\title{CAT configuration}
\description{
An S4 Class to represent parameters of the CAT configuration.
}
\section{Slots}{

\describe{
\item{\code{solverConfig}}{an instance of the S4 class \code{SolverConfig} for the 
MIP solver configuration.}

\item{\code{initialTheta}}{the initial ability theta value.}

\item{\code{scalingConstant}}{the constraint to scale a discrimination coefficient.
estimated with the logistic model to the normal metric.}

\item{\code{scoreMethodConfig}}{a rJava \code{jobjRef} object for CAT scoring method
configuration. It is generated by the function \code{scoreMethodConfig}.}

\item{\code{exposureControlType}}{a character string specifying the exposure control
type. "None" to disable exposure control, "Item" for item-level exposure 
control, and "Passage" for passage-level exposure control.}

\item{\code{exposureControlRate}}{an integer value specifying the exposure goal rate.}

\item{\code{lValue}}{a non-negative integer specifying the nuumber of items to be
randomlized.}
}}

