% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EAPConfig.R
\docType{class}
\name{EAPConfig-class}
\alias{EAPConfig-class}
\alias{EAPConfig}
\title{EAP configuration}
\description{
An S4 class to represent expected A posteriori (EAP) scoring algorithm 
configuration.
}
\details{
An instance of this S4 class can be applied to the generic function 
\code{scoreMethodConfig} to create an Java object for scoring method 
configuration.
}
\section{Slots}{

\describe{
\item{\code{numQuad}}{a positive integer specifying the number of quadrature points}

\item{\code{minQuad}}{a numeric value specifying the minimum quadrature point}

\item{\code{maxQuad}}{a numeric value specifying the maximum quadrature point}

\item{\code{priorDistType}}{a character string specifying the prior distribution of 
ability. "Normal" for Normal distribution and "Uniform" for uniform 
distribution.}

\item{\code{distParams}}{a numeric vector specifying parameters of the prior
distribution. (mean, sd) for the Normal distribution, (a, b) for the 
uniform distribution.}
}}

