% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launchApp.R
\name{launchApp}
\alias{launchApp}
\title{Launches the shiny app to confiure and run CAT simulations.}
\usage{
launchApp()
}
\description{
Launches the shiny app to confiure and run CAT simulations.
}
\examples{
if(interactive()){
launchApp()
} 
}
