% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilFunctions.R
\name{result2CSV}
\alias{result2CSV}
\title{Creates a simulation result CSV file.}
\usage{
result2CSV(simResult, file)
}
\arguments{
\item{simResult}{an instance of S4 class \code{SimResult}.}

\item{file}{a writable connection or a character string naming the file 
to write to.}
}
\description{
Creates a simulation result CSV file.
}
