% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoreMethodConfig.R
\name{scoreMethodConfig}
\alias{scoreMethodConfig}
\alias{scoreMethodConfig,EAPConfig-method}
\title{Creates a scoring method configuration for CAT simulation}
\usage{
scoreMethodConfig(object)

\S4method{scoreMethodConfig}{EAPConfig}(object)
}
\arguments{
\item{object}{an S4 object for the estimation algorithm configuration}
}
\value{
the object of scoring method configuration which is an instance of
  \code{org/act/cat/ScoringMethodConfig}
}
\description{
This is a gneric function to create a scoring method configuration from a
speicfic estimation algorithm configuration.
}
\examples{
if(interactive()){
eapConfig <- EAPConfig(numQuad = 6L, minQuad = -2, maxQuad = 2, 
  priorDistType = "Normal", distParams = c(0,1))
scoreMethodConfig <- scoreMethodConfig(eapConfig)  
}
}
