% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setupJars.R
\name{setupJars}
\alias{setupJars}
\title{Sets up JAR dependency.}
\usage{
setupJars()
}
\description{
This function downloads JAR dependencies securely through https and put them in
/java
}
\examples{
\dontrun{
setupJars()
}
}
