% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setupXprm.R
\name{setupXprm}
\alias{setupXprm}
\title{Sets up the required JAR dependancy to parse Mosel.}
\usage{
setupXprm(path)
}
\arguments{
\item{path}{A character string specifying the xprm.jar file in the installed 
Xpress folder.}
}
\description{
Sets up the required JAR dependancy to parse Mosel.
}
\examples{
\dontrun{
setupXprm(C:/xpressmp/lib/xprm.jar)
} 
}
