% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyAppServer.R
\name{shinyAppServer}
\alias{shinyAppServer}
\title{Defines server logic to configure and run CAT simulations.}
\usage{
shinyAppServer(input, output)
}
\arguments{
\item{input}{an object that stores the current values of all of the widgets 
in the app.}

\item{output}{an object that stores instructions for building the R objects
in the app.}
}
\description{
Defines server logic to configure and run CAT simulations.
}
