% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configClasses.R
\docType{class}
\name{SimConfig-class}
\alias{SimConfig-class}
\alias{SimConfig}
\title{CAT simulation configuration}
\description{
An S4 class to represent CAT simulation configuraiton.
}
\section{Slots}{

\describe{
\item{\code{simID}}{a character string as the identifier of the CAT simulation.}

\item{\code{numExaminees}}{a positive integer specifying the number of simulated 
examinees.}

\item{\code{trueThetaDistType}}{a character string specifying the distribution of 
true ability of simulated examinees. "Normal" for the Normal distribution 
and "Uniform" for the uniform distribution.}

\item{\code{trueThetaDistParams}}{a numeric vector spcifying parameters of the prior
distribuution. (mean, sd) for the Normal distribution, (a, b) for the 
uniform distribution.}
}}

