% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configClasses.R
\docType{class}
\name{SolverConfig-class}
\alias{SolverConfig-class}
\alias{SolverConfig}
\title{MIP solver configuration}
\description{
An S4 Class to represent parameters of the MIP solver configuration.
}
\section{Slots}{

\describe{
\item{\code{absGap}}{the absolute gap target to terminate the MIP solving.}

\item{\code{relGap}}{the relative gap target to terminate the MIP solving.}

\item{\code{intTol}}{the integer tolerance for the MIP solving. if the solution x 
is between -intTol and intTol, x <= 0 is true if the value of x is at most 
intTol. x > 0 is fulfilled if x > intTol.}
}}

