% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilFunctions.R
\name{summary,SimResult-method}
\alias{summary,SimResult-method}
\title{Generates CAT simulation summary}
\usage{
\S4method{summary}{SimResult}(object)
}
\arguments{
\item{object}{an object of \code{SimResult}.
Generates the summary report of CAT simulation.}
}
\description{
Generates CAT simulation summary
}
