% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configClasses.R
\docType{class}
\name{TestConfig-class}
\alias{TestConfig-class}
\alias{TestConfig}
\title{Test specification configuration}
\description{
An S4 class to represent test blueprint and specification.
}
\section{Slots}{

\describe{
\item{\code{testConfigID}}{a character string as the identifier of the test 
configuration.}

\item{\code{testLength}}{a positive integer specifying the test length.}

\item{\code{itempoolPath}}{a character string specifying the location of the item pool
csv file.}

\item{\code{passagepoolPath}}{a character string specifying the location of the 
passage pool csv file.}

\item{\code{constraintPath}}{a character string specifying the location of the
constraint csv file.}

\item{\code{itemNumericColumn}}{a boolean vector indicating whether item attribute 
columns in the item pool table are numeric or not.}

\item{\code{passageNumericColumn}}{a boolean vector indicating whether passage
attribute columns in the passage pool table are numeric or not.}

\item{\code{enableEnemyItem}}{a boolean indicator to specify if enemy item 
constraints defined by in the item pool is enabled or not.}

\item{\code{numPassageLB}}{an integer specifying the minimum number of passages in 
the test.}

\item{\code{numPassageUB}}{an integer specifying the maximum number of passages in 
the test.}

\item{\code{numItemPerPassageLB}}{an integer specifying the minimum number of items
in a passages in the test.}

\item{\code{numItemPerPassageUB}}{an integer specifying the maximum number of items
in a passages in the test.}
}}

