% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setupJars.R
\name{setupJars}
\alias{setupJars}
\title{Setups JAR dependency.}
\usage{
setupJars()
}
\description{
This function downloads JAR dependency from mvnrepository and put them in
/java
}
\examples{
\dontrun{
setupJars()
}
}
